chore(cra): Downgrade `react-scripts`
`craco` doesn't yet support v5, which is causing the `npm` branch to
throw errors when installing. While `yarn` doesn't throw, it's better to
keep both branches in sync (will have to silence renovatebot when it
tries to upgrade it).

Moreover, react-scripts@5 has a lot of peer dependency warnings in yarn.